refactor tags, study card
