+++ client/src/views/Mentor/Classroom/Classroom.jsx
@@ -7,6 +7,7 @@ import Roster from './Roster/Roster';
 import Home from './Home/Home';
 import SavedWorkSpaceTab from '../../../components/Tabs/SavedWorkspaceTab';
 import { useSearchParams, useParams } from 'react-router-dom';
+import React from 'react';
 
 const { TabPane } = Tabs;
